stack: aptx adaptive encoder
Changing sampling rate from 16 to 24.
Allocating memory using size_t in osi_calloc
Inform mode info to SoC.

+++ bta/av/bta_av_api.cc
@@ -747,7 +747,7 @@ void BTA_AvUpdateCodecSupport(uint8_t *p_codec_type_list, uint8_t *p_vnd_list, u
     if (p_codec_type_list == NULL || p_vnd_list == NULL || p_codec_id_list == NULL)
         return;
 
-    if ((p_buf = (tBTA_AV_UPDATE_SUPP_CODECS *) osi_calloc((uint8_t) (sizeof
+    if ((p_buf = (tBTA_AV_UPDATE_SUPP_CODECS *) osi_calloc((sizeof
         (tBTA_AV_UPDATE_SUPP_CODECS)))) != NULL)
     {
         p_buf->hdr.event = BTA_AV_UPDATE_SUPP_CODECS;
